Split sync task into bundle and resource sync tasks
No functional changes
